Added argument parsing

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: malmeida <malmeida@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 12:33:08 by malmeida          #+#    #+#             */
-/*   Updated: 2021/08/31 13:58:38 by malmeida         ###   ########.fr       */
+/*   Updated: 2021/09/20 23:05:57 by malmeida         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ char				*ft_strrchr(const char *s, int c);
 char				*ft_strnstr(const char *haystack, const char *needle,
 						size_t len);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
-int					ft_atoi(const char *str);
+long int			ft_atoi(const char *str);
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
 int					ft_isalnum(int c);
